New HLA conversion interface and code restructure
