+++ pom.xml
@@ -59,6 +59,11 @@
 			<artifactId>ashley</artifactId>
 			<version>1.7.3</version>
 		</dependency>
+		<dependency>
+			<groupId>org.apache.commons</groupId>
+			<artifactId>commons-collections4</artifactId>
+			<version>4.5.0-M3</version>
+		</dependency>
 	</dependencies>
 
 	<build>
