+++ settings/HlaConfig.xml
@@ -1,0 +1,26 @@
+<Project>
+	<RtiConfiguration Host="localhost" Port="8989" Federation="SEE 2025" FederateType="Outpost">
+		<FomModules>
+			<Fom FilePath="foms/SISO_SpaceFOM_datatypes.xml" />
+			<Fom FilePath="foms/SISO_SpaceFOM_entity.xml" />
+			<Fom FilePath="foms/SISO_SpaceFOM_environment.xml" />
+			<Fom FilePath="foms/SISO_SpaceFOM_management.xml" />
+			<Fom FilePath="foms/SISO_SpaceFOM_switches.xml" />
+		</FomModules>
+	</RtiConfiguration>
+	<Simulation FrameRate="60" />
+	<Entities>
+		<Entity Class="HLAobjectRoot.PhysicalEntity">
+			<Attribute Name="name" Publish="True" Subscribe="True" />
+			<Attribute Name="parent_reference_frame" Publish="True" Subscribe="True" />
+			<Attribute Name="state" Publish="True" Subscribe="True" />
+			<Attribute Name="center_of_mass" Publish="True" Subscribe="False" />
+			<Attribute Name="acceleration" Publish="True" Subscribe="False" />
+		</Entity>
+	</Entities>
+	<Interactions>
+		<Interaction Class="HLAinteractionRoot.ModeTransitionRequest" Publish="True" Subscribe="True">
+			<Parameter Name="execution_mode" />
+		</Interaction>
+	</Interactions>
+</Project>
