+++ src/main/java/atreia108/vega/core/IWorld.java
@@ -1,0 +1,8 @@
+package atreia108.vega.core;
+
+import com.badlogic.ashley.core.Entity;
+
+public interface IWorld {
+	public Entity createEntity();
+	public void destroyEntity();
+}
