+++ src/main/java/atreia108/vega/types/IComponent.java
@@ -1,0 +1,5 @@
+package atreia108.vega.types;
+
+import com.badlogic.ashley.core.Component;
+
+public interface IComponent extends Component {}
