--- src/main/java/atreia108/vega/utils/HlaInteractionClass.java
@@ -1,67 +1,0 @@
-/*-
- * SPDX-License-Identifier: BSD-3-Clause
- * Copyright (c) 2025 Hridyanshu Aatreya <2200096@brunel.ac.uk>
- * All rights reserved.
-
- * Redistribution and use in source and binary forms, with or without 
- * modification, are permitted provided that the following conditions 
- * are met:
- * 1. Redistributions of source code must retain the above copyright notice, 
- *	  this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the 
- *    documentation and/or other materials provided with the distribution.
- * 3. Neither the name of the copyright holder nor the names of its 
- * 	  contributors may be used to endorse or promote products derived from this
- *    software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS ``AS IS'' AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED.  IN NO EVENT SHALL THE AUTHOR OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
- * OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
- * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
- * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
- * SUCH DAMAGE.
- * 
- */
-
-package atreia108.vega.utils;
-
-import java.util.HashSet;
-import java.util.Set;
-
-public class HlaInteractionClass {
-	private String className;
-	HlaMessagePattern messagePattern;
-	private Set<String> parameterSet;
-	
-	public HlaInteractionClass(String className, HlaMessagePattern messagePattern) {
-		parameterSet = new HashSet<String>();
-		this.className = className;
-		this.messagePattern = messagePattern;
-	}
-	
-	public String getName() { return className; }
-	
-	public void registerParameter(String parameterName) {
-		parameterSet.add(parameterName);
-	}
-	
-	public boolean isPublishable() {
-		if (messagePattern == HlaMessagePattern.PUBLISH_ONLY || messagePattern == HlaMessagePattern.PUBLISH_SUBSCRIBE)
-			return true;
-		else return false;
-	}
-	
-	public boolean isSubscribeable() {
-		if (messagePattern == HlaMessagePattern.SUBSCRIBE_ONLY || messagePattern == HlaMessagePattern.PUBLISH_SUBSCRIBE)
-			return true;
-		else return false;
-	}
-	
-	public Set<String> getParameterSet() { return parameterSet; }
-}
